+++ springboot/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 
 spring.datasource.url=jdbc:postgresql://localhost:5432/posBD
 spring.datasource.username=postgres
-spring.datasource.password=admin
+spring.datasource.password=123
 spring.datasource.driverClassName=org.postgresql.Driver
 
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults = false
